+++ Cli/Command/EntityFromTable.php
@@ -61,7 +61,7 @@ class EntityFromTable extends Command
      *
      * @return int
      *
-     * @throws \Exception
+     * @throws \Symfony\Component\Console\Exception\ExceptionInterface
      */
     protected function execute(InputInterface $input, OutputInterface $output) : int
     {