+++ Finder/EmailLog.php
@@ -4,6 +4,11 @@ namespace TickTackk\DeveloperTools\Finder;
 
 use XF\Mvc\Entity\Finder;
 
+/**
+ * @method \XF\Mvc\Entity\AbstractCollection<\TickTackk\DeveloperTools\Entity\EmailLog> fetch($limit = null, $offset = null)
+ * @method \TickTackk\DeveloperTools\Entity\EmailLog|null fetchOne($limit = null, $offset = null)
+ * @extends \XF\Mvc\Entity\Finder<\TickTackk\DeveloperTools\Entity\EmailLog>
+ */
 class EmailLog extends Finder
 {
 }