+++ XF/Mail/XF2/Mailer.php
@@ -17,7 +17,8 @@ class Mailer extends XFCP_Mailer
      *
      * @throws \XF\PrintableException
      *
-     * @noinspection PhpHierarchyChecksInspection
+     * @noinspection PhpUndefinedClassInspection
+     * @noinspection PhpSignatureMismatchDuringInheritanceInspection
      */
     public function send
     (